get rid of aclk_api.[ch]
