+++ Makefile.am
@@ -860,8 +860,6 @@ endif #ENABLE_ACLK
 
 ACLK_ALWAYS_BUILD_FILES = \
     aclk/aclk_rrdhost_state.h \
-    aclk/aclk_api.c \
-    aclk/aclk_api.h \
     aclk/aclk_proxy.c \
     aclk/aclk_proxy.h \
     $(NULL)
