+++ aclk/aclk.h
@@ -10,6 +10,17 @@
 // stable for the purposes of TBEB (truncated binary exponential backoff)
 #define ACLK_PUBACKS_CONN_STABLE 3
 
+extern int aclk_connected;
+extern int aclk_alert_reloaded;
+extern int use_mqtt_5;
+extern int aclk_ctx_based;
+extern int aclk_disable_runtime;
+extern int aclk_stats_enabled;
+extern int aclk_kill_link;
+
+extern usec_t aclk_session_us;
+extern time_t aclk_session_sec;
+
 extern time_t aclk_block_until;
 
 extern int disconnect_req;
@@ -34,7 +45,8 @@ void aclk_send_node_instances(void);
 
 void aclk_send_bin_msg(char *msg, size_t msg_len, enum aclk_topics subtopic, const char *msgname);
 
-char *ng_aclk_state(void);
-char *ng_aclk_state_json(void);
+char *aclk_state(void);
+char *aclk_state_json(void);
+void add_aclk_host_labels(void);
 
 #endif /* ACLK_H */
