+++ aclk/aclk_util.c
@@ -1,6 +1,7 @@
 // SPDX-License-Identifier: GPL-3.0-or-later
 
 #include "aclk_util.h"
+#include "aclk_proxy.h"
 
 #include "daemon/common.h"
